Address cli parameter parsing issue
When omitting an attribute level from the commandline, the script tries to
interpret whether there is no level or specific attribute name to search for,
or whether the operator put an attribute name in instead of a level. This
process was incorrectly using a symbol for the attribute name and trying to
look the name up at level :all, which ended up failing.

This change avoids the symbol convert and alters the code to look up at
:default by default.